requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-chess==1.10.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
